pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "novel-notify"
version = "0.1.0"
description = "A Telegram bot for novel update notifications from WebNovel"
authors = [
    {name = "Invictus", email = "invictus@example.com"},
]
dependencies = [
    "python-telegram-bot[webhooks]==22.1",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "GPL-3.0"}

[project.scripts]
novel-notify = "novel_notify.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
